Added test structure for multiply_int16
