+++ test/unit_test/src/main.c
@@ -35,5 +35,7 @@ int main(void) {
 
   CALL(test_expand_8_to_16);
   CALL(test_output_transform_16);
+  CALL(test_multiply_int16);
+
   return UNITY_END();
 }
